Fix test, mostly due to decimal point and accidental of to_i for payment_amount

+++ lib/payments_factory.rb
@@ -11,7 +11,7 @@ class PaymentsFactory < Factory
 
       parse(json) do |payment|
         user = payment['user']
-        amount = payment['amount'].to_i
+        amount = payment['amount'].to_f
 
         payments[user] = [] if payments[user].nil?
         payments[user] << Payment.new(
